+++ qt/network/gen_qtcpserver.go
@@ -126,12 +126,12 @@ func (this *QTcpServer) ServerAddress() *QHostAddress {
 	return _goptr
 }
 
-func (this *QTcpServer) SocketDescriptor() uintptr {
-	return (uintptr)(C.QTcpServer_SocketDescriptor(this.h))
+func (this *QTcpServer) SocketDescriptor() int64 {
+	return (int64)(C.QTcpServer_SocketDescriptor(this.h))
 }
 
-func (this *QTcpServer) SetSocketDescriptor(socketDescriptor uintptr) bool {
-	return (bool)(C.QTcpServer_SetSocketDescriptor(this.h, (C.intptr_t)(socketDescriptor)))
+func (this *QTcpServer) SetSocketDescriptor(socketDescriptor int64) bool {
+	return (bool)(C.QTcpServer_SetSocketDescriptor(this.h, (C.longlong)(socketDescriptor)))
 }
 
 func (this *QTcpServer) WaitForNewConnection() bool {
@@ -322,12 +322,12 @@ func miqt_exec_callback_QTcpServer_NextPendingConnection(self *C.QTcpServer, cb 
 
 }
 
-func (this *QTcpServer) callVirtualBase_IncomingConnection(handle uintptr) {
+func (this *QTcpServer) callVirtualBase_IncomingConnection(handle int64) {
 
-	C.QTcpServer_virtualbase_IncomingConnection(unsafe.Pointer(this.h), (C.intptr_t)(handle))
+	C.QTcpServer_virtualbase_IncomingConnection(unsafe.Pointer(this.h), (C.longlong)(handle))
 
 }
-func (this *QTcpServer) OnIncomingConnection(slot func(super func(handle uintptr), handle uintptr)) {
+func (this *QTcpServer) OnIncomingConnection(slot func(super func(handle int64), handle int64)) {
 	if !this.isSubclass {
 		panic("miqt: can only override virtual methods for directly constructed types")
 	}
@@ -335,14 +335,14 @@ func (this *QTcpServer) OnIncomingConnection(slot func(super func(handle uintptr
 }
 
 //export miqt_exec_callback_QTcpServer_IncomingConnection
-func miqt_exec_callback_QTcpServer_IncomingConnection(self *C.QTcpServer, cb C.intptr_t, handle C.intptr_t) {
-	gofunc, ok := cgo.Handle(cb).Value().(func(super func(handle uintptr), handle uintptr))
+func miqt_exec_callback_QTcpServer_IncomingConnection(self *C.QTcpServer, cb C.intptr_t, handle C.longlong) {
+	gofunc, ok := cgo.Handle(cb).Value().(func(super func(handle int64), handle int64))
 	if !ok {
 		panic("miqt: callback of non-callback type (heap corruption?)")
 	}
 
 	// Convert all CABI parameters to Go parameters
-	slotval1 := (uintptr)(handle)
+	slotval1 := (int64)(handle)
 
 	gofunc((&QTcpServer{h: self}).callVirtualBase_IncomingConnection, slotval1)
